+++ fuentes/BACK/backend/src/main/java/trainer/api/backend/model/dao/IObjetivoDao.java
@@ -9,5 +9,5 @@ import java.util.List;
 
 public interface IObjetivoDao extends CrudRepository<Objetivo, Long> {
     @Query("FROM Objetivo o WHERE o.usuarioId=:id")
-    List<Objetivo> findListByUserId(@Param("id") Long id);
+    List<Objetivo> findListByUserId(@Param("id") Integer id);
 }
